fix: Correct recipe rating logic and remove redundant spacing in HomeViewModel
- Updated rating calculation in `RecipeCard` to use `toDouble()` for accuracy.
- Removed unnecessary blank lines in `HomeViewModel` for improved readability.

+++ app/src/main/java/com/example/cookat/screens/home/components/RecipeCard.kt
@@ -101,7 +101,7 @@ fun RecipeCard(
 					) {
 						Row {
 							repeat(5) { i ->
-								val starFilled = i < recipe.rating.toInt()
+								val starFilled = i < recipe.rating.toDouble()
 								Icon(
 									imageVector = if (starFilled) Icons.Filled.Star else Icons.Filled.StarBorder,
 									contentDescription = null,
